requirements.txt
google-genai
pytrends
feedparser
requests
python-dotenv
beautifulsoup4
google-play-scraper
